+++ octodns/__init__.py
@@ -1,10 +1,3 @@
 'OctoDNS: DNS as code - Tools for managing DNS across multiple providers'
 
-from __future__ import (
-    absolute_import,
-    division,
-    print_function,
-    unicode_literals,
-)
-
 __VERSION__ = '0.9.19'