+++ tests/test_octodns_manager.py
@@ -2,13 +2,6 @@
 #
 #
 
-from __future__ import (
-    absolute_import,
-    division,
-    print_function,
-    unicode_literals,
-)
-
 from os import environ
 from os.path import dirname, isfile, join