+++ CHANGELOG.md
@@ -12,6 +12,8 @@
   Provider.supports and just always says Yes so that any dynamically registered
   types will be supported.
 * Add TtlRestrictionFilter processor for adding ttl restriction/checking
+* NameAllowlistFilter & NameRejectlistFilter implementations to support
+  filtering on record names to include/exclude records from management.
 
 ## v0.9.18 - 2022-08-14 - Subzone handling